backend/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
subprocess
